feat(trips): add `Trip`, `TripRepository` and `CreateTripService` (#31)
* feat: add `Entity` generic

* chore(deps): add `nanoid`

* feat: add `Trip` entity (#31)

* refactor: add `Controller` generic

* refactor: boilerplate `create-trip` feature

* refactor: discard idea of `Controller`

* docs: describe `Entity` idea

* refactor: add `Usecase` generic

* refactor: implement bare `CreateTripController`

* feat: add `TripMapper`

* feat: add `TripRepository`

* fix: add missing `DatabaseRecords`

* fix: add missing `Repository<T>`

* refactor: add `TripModule`

* refactor: remove `AppController` and `AppService`

* refactor: improve dependency injection

* chore: add swagger documentation (#23)

* feat: boilerpalte `create-trip` functionality

* refactor: change currency display

* refactor: move `CreateTripResponse` and `CreateTripRequest`

* feat: implement `TripRepository` and `TripMapper`

* refactor: use relative imports

* test: add `create-trip` tests

* refactor: update examples of `CreateTripResponse`

* ci: add `postgres` service to pipeline

* ci: rename `DATABASE_URL` to `DATABASE_URI`

* refactor: remove console.log used for debugging

* ci: prepare development database

* build: add prisma to docker image

* build: create `/prisma` directory

* build: fix wrong path for prisma

* build: remove useless `mkdir` step

+++ src/modules/geocoding/services/geocode-maps/geocode-maps.geocoding.ts
@@ -31,6 +31,6 @@ export class GeocodeMapsGeocodingService extends GeocodingService {
       throw new Error('No coordinates found')
     }
 
-    return new Coordinates(Number.parseInt(lat), Number.parseInt(lon))
+    return new Coordinates(Number.parseFloat(lat), Number.parseFloat(lon))
   }
 }
